speed_test_gui/requirements.txt

speedtest-cli
